+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Next version
 
+- Tasks : add :on_warning_or_alert scope/filter
+- Tasks : fix redirect after create comment
+
 ## 3.9.5
 
 - Fix BillingMachine preview JS